chore: zk.js consistent file and folder naming and structure
