+++ zk.js/src/index.ts
@@ -1,18 +1,18 @@
-export * from "./utxo";
-export * from "./programUtxo";
-export * from "./test-utils";
 export * from "./account";
+// export * from "./balance"
+export * from "./build-balance";
 export * from "./constants";
-export * from "./wallet";
-export * from "./utils";
-export * from "./idls";
-export * from "./rpc";
-export * from "./merkleTree";
 export * from "./errors";
+export * from "./provider";
+export * from "./rpc";
+
+export * from "./test-utils";
+export * from "./wallet";
+export * from "./idls";
+export * from "./merkle-tree";
 export * from "./types";
 export * from "./transaction";
-export * from "./convertCase";
-export * from "./closeVerifierState";
-export * from "./createRustVerifyingKey";
+export * from "./utils";
+export * from "./utxo";
 
 export * as circuitlibjs from "@lightprotocol/circuit-lib.js";
