+++ zk.js/src/merkle-tree/index.ts
@@ -1,0 +1,2 @@
+export * from "./sol-merkle-tree";
+export * from "./merkle-tree-config";
