--- zk.js/src/merkleTree/index.ts
@@ -1,2 +1,0 @@
-export * from "./solMerkleTree";
-export * from "./merkleTreeConfig";
