+++ zk.js/src/test-utils/functional-circuit.ts
