+++ zk.js/src/test-utils/init-lookuptable.ts
