+++ zk.js/src/utxo/index.ts
@@ -1,0 +1,4 @@
+export * from "./create-output-utxos";
+export * from "./program-utxo";
+export * from "./select-input-utxos";
+export * from "./utxo";
