+++ zk.js/src/wallet/index.ts
@@ -1,6 +1,6 @@
-export * from "./buildBalance";
+export * from "../build-balance";
 export * from "./user";
-export * from "./provider";
-export * from "./createOutUtxos";
-export * from "./selectInUtxos";
-export * from "./useWallet";
+export * from "../provider";
+export * from "../utxo/create-output-utxos";
+export * from "../utxo/select-input-utxos";
+export * from "./use-wallet";
